#18 Move utility modules to util
