+++ src/commands/miscellaneous/ping.ts
@@ -23,8 +23,8 @@
  */
 
 import { SlashCommandBuilder, EmbedBuilder, ChatInputCommandInteraction } from "discord.js";
-import { Command } from "../../command-template.js";
-import { defaultErrorHandler } from "../../error-handler.js";
+import { Command } from "../../util/command-template.js";
+import { defaultErrorHandler } from "../../util/error-handler.js";
 
 /*
  * Displays the bot's latency in milliseconds
